@@ -1,11 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import { AuthContextProvider } from './context/authContext/AuthContext'
-import { MovieContextProvider } from './context/movieContext/MovieContext';
+import ReactDOM from 'react-dom/client';
+import App from './App.jsx';
+import { AuthContextProvider } from './context/authContext/AuthContext';
 import { ListContextProvider } from './context/listContext/ListContext';
+import { MovieContextProvider } from './context/movieContext/MovieContext';
+import './index.css';
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <AuthContextProvider>
       <MovieContextProvider>
@@ -15,5 +16,4 @@ ReactDOM.render(
       </MovieContextProvider>
     </AuthContextProvider>
   </React.StrictMode>,
-  document.getElementById('root')
-);
+)
