@@ -4,7 +4,7 @@ import { loginFailure, loginStart, loginSuccess } from "./AuthActions";
 export const login = async (user, dispatch) => {
     dispatch(loginStart());
     try {
-        const res = await axios.post("auth/login", user);
+        const res = await axios.post("/api/auth/login", user);
         res.data.isAdmin && dispatch(loginSuccess(res.data));
         console.log(res.data)
     } catch {
