@@ -4,7 +4,7 @@ import axios from 'axios';
 export const getLists = async (dispatch) => {
     dispatch(getListsStart())
     try {
-        const res = await axios.get("/lists", {
+        const res = await axios.get("/api/lists", {
             headers: { token: "Bearer " + JSON.parse(localStorage.getItem("user")).accessToken }
         });
         dispatch(getListsSuccess(res.data))
@@ -17,7 +17,7 @@ export const getLists = async (dispatch) => {
 export const deleteList = async (id, dispatch) => {
     dispatch(deleteListStart());
     try {
-        await axios.delete("/lists/" + id, {
+        await axios.delete("/api/lists/" + id, {
             headers: { token: "Bearer " + JSON.parse(localStorage.getItem("user")).accessToken }
         });
         dispatch(deleteListSuccess(id))
@@ -30,7 +30,7 @@ export const deleteList = async (id, dispatch) => {
 // export const updatelist = async (list, dispatch) => {
 //     dispatch(updatelistStart());
 //     try {
-//         const res = await axios.put("/lists/" + list._id,list, {
+//         const res = await axios.put("/api/lists/" + list._id,list, {
 //             headers: { token: "Bearer " + JSON.parse(localStorage.getItem("user")).accessToken }
 //         });
 //         dispatch(updatelistSuccess(res.data))
@@ -44,7 +44,7 @@ export const createList = async (list, dispatch) => {
     dispatch(createListStart());
     try {
         
-        const res =await axios.post("/lists", list, {
+        const res =await axios.post("/api/lists", list, {
             headers: { token: "Bearer " + JSON.parse(localStorage.getItem("user")).accessToken }
         });
         dispatch(createListSuccess(res.data));
